replace rps with better rps
